+++ R/derive_param_extreme_event.R
@@ -60,6 +60,13 @@
 #'
 #'   *Permitted Values*: `"first"`, `"last"`
 #'
+#' @param keep_vars_source Variables to be kept in the new records
+#'
+#'   A named list returned by `exprs()` defining the variables to be kept for the
+#'   new records.
+#'
+#'   *Default*: `everything()`
+#'
 #' @param set_values_to Variables to set
 #'
 #'   A named list returned by `exprs()` defining the variables to be set for the
@@ -113,6 +120,7 @@
 #' library(lubridate)
 #'
 #' # Derive a new parameter for the first disease progression (PD)
+#' # Specify the variables that need to be kept in the new records.
 #' adsl <- tribble(
 #'   ~USUBJID, ~DTHDT,
 #'   "1",      ymd("2022-05-13"),
@@ -122,14 +130,14 @@
 #'   mutate(STUDYID = "XX1234")
 #'
 #' adrs <- tribble(
-#'   ~USUBJID, ~ADTC,        ~AVALC,
-#'   "1",      "2020-01-02", "PR",
-#'   "1",      "2020-02-01", "CR",
-#'   "1",      "2020-03-01", "CR",
-#'   "1",      "2020-04-01", "SD",
-#'   "2",      "2021-06-15", "SD",
-#'   "2",      "2021-07-16", "PD",
-#'   "2",      "2021-09-14", "PD"
+#'   ~USUBJID, ~RSSEQ, ~ADTC,        ~AVALC,
+#'   "1",      1,      "2020-01-02", "PR",
+#'   "1",      2,      "2020-02-01", "CR",
+#'   "1",      3,      "2020-03-01", "CR",
+#'   "1",      4,      "2020-04-01", "SD",
+#'   "2",      1,      "2021-06-15", "SD",
+#'   "2",      2,      "2021-07-16", "PD",
+#'   "2",      3,      "2021-09-14", "PD"
 #' ) %>%
 #'   mutate(
 #'     STUDYID = "XX1234",
@@ -150,11 +158,11 @@
 #'   true_value = "Y",
 #'   false_value = "N",
 #'   mode = "first",
+#'   keep_vars_source = exprs(USUBJID, STUDYID, ADT),
 #'   set_values_to = exprs(
 #'     PARAMCD = "PD",
 #'     PARAM = "Disease Progression",
-#'     ANL01FL = "Y",
-#'     ADT = ADT
+#'     ANL01FL = "Y"
 #'   )
 #' )
 #'
@@ -167,6 +175,7 @@
 #'   true_value = "Y",
 #'   false_value = "N",
 #'   mode = "first",
+#'   keep_vars_source = NULL,
 #'   set_values_to = exprs(
 #'     PARAMCD = "DEATH",
 #'     PARAM = "Death",
@@ -184,12 +193,22 @@ derive_param_extreme_event <- function(dataset = NULL,
                                        false_value = "N",
                                        mode = "first",
                                        subject_keys = get_admiral_option("subject_keys"),
+                                       keep_vars_source = NULL,
                                        set_values_to,
                                        check_type = "warning") {
   # Check input parameters
   filter_source <- assert_filter_cond(enexpr(filter_source))
   assert_vars(subject_keys)
   assert_vars(order, optional = TRUE)
+  assert_vars(keep_vars_source, optional = TRUE)
+
+
+  if (!is.null(keep_vars_source)) {
+    keep_vars_source <- keep_vars_source
+  } else {
+    keep_vars_source <- exprs(everything())
+  }
+
   assert_data_frame(dataset_source,
     required_vars = exprs(!!!subject_keys, !!!extract_vars(order))
   )
@@ -238,6 +257,7 @@ derive_param_extreme_event <- function(dataset = NULL,
   }
 
   new_obs <- bind_rows(events, noevents) %>%
+    select(!!!keep_vars_source, !!new_var) %>%
     mutate(
       !!!set_values_to
     )